Pc 9088 train model with mlops (#306)
* (PC-9016): refresh v1 recommendations

* (PC-9016): update tests

* (PC-9088): clean data collection step

* (PC-9088): clean dag

* (PC-9088): clean preprocess step

* (PC-9088): clean training step

* (PC-9088): clean evaluation step

* (PC-9088): clean evaluation step - black

* (PC-9088): debug scripts

* (PC-9088): debug evaluate

* (PC-9088): split triplet model code

* (PC-9088): rename variables

* (PC-9088): debug model

* (PC-9088): log parameters

* (PC-9088): last 10 days data and dev branch for testing

* (PC-9088): debug log_param

* (PC-9088): remove test params
