+++ algo_training/models/margin_loss.py
@@ -8,8 +8,8 @@ class MarginLoss(layers.Layer):
         self.margin = margin
 
     def call(self, inputs):
-        pos_pair_similarity = inputs[0]
-        neg_pair_similarity = inputs[1]
+        positive_pair_similarity = inputs[0]
+        negative_pair_similarity = inputs[1]
 
-        diff = neg_pair_similarity - pos_pair_similarity
+        diff = negative_pair_similarity - positive_pair_similarity
         return tf.maximum(diff + self.margin, 0.0)
