+++ algo_training/models/metrics.py
@@ -1,0 +1,195 @@
+import numpy as np
+import tqdm
+from scipy.spatial.distance import cosine
+
+
+TYPE_LIST = [
+    "ThingType.LIVRE_EDITION",
+    "ThingType.INSTRUMENT",
+    "EventType.PRATIQUE_ARTISTIQUE",
+    "ThingType.AUDIOVISUEL",
+    "ThingType.MUSIQUE",
+    "EventType.SPECTACLE_VIVANT",
+    "ThingType.PRATIQUE_ARTISTIQUE_ABO",
+    "EventType.CINEMA",
+    "EventType.MUSIQUE",
+    "EventType.CONFERENCE_DEBAT_DEDICACE",
+    "EventType.MUSEES_PATRIMOINE",
+    "ThingType.MUSEES_PATRIMOINE_ABO",
+    "ThingType.PRESSE_ABO",
+    "ThingType.OEUVRE_ART",
+    "ThingType.LIVRE_AUDIO",
+    "EventType.JEUX",
+    "ThingType.CINEMA_ABO",
+    "ThingType.JEUX_VIDEO",
+    "ThingType.CINEMA_CARD",
+    "ThingType.MUSIQUE_ABO",
+    "EventType.ACTIVATION",
+    "ThingType.JEUX_VIDEO_ABO",
+    "ThingType.SPECTACLE_VIVANT_ABO",
+    "ThingType.ACTIVATION",
+]
+
+
+def get_unexpectedness(booked_type_list, recommended_type_list):
+    booked_type_vector_list = [
+        [int(booked_type == offer_type) for offer_type in TYPE_LIST]
+        for booked_type in booked_type_list
+    ]
+    recommended_type_vector_list = [
+        [int(recommended_type == offer_type) for offer_type in TYPE_LIST]
+        for recommended_type in recommended_type_list
+    ]
+
+    cosine_sum = 0
+    for booked_type_vector in booked_type_vector_list:
+        for recommended_type_vector in recommended_type_vector_list:
+            cosine_sum += cosine(booked_type_vector, recommended_type_vector)
+
+    return (1 / (len(booked_type_list) * len(recommended_type_list))) * cosine_sum
+
+
+def compute_metrics(k, positive_data_train, positive_data_test, match_model):
+    # Map all offers to corresponding types
+    offer_type_dict = {}
+    unique_offer_types = (
+        positive_data_train.groupby(["item_id", "type"]).first().reset_index()
+    )
+    for item_id, item_type in zip(
+        unique_offer_types.item_id.values, unique_offer_types.type.values
+    ):
+        offer_type_dict[item_id] = item_type
+
+    # Only keep user - item iteractions in positive_data_test, which can be infered from model
+    cleaned_positive_data_test = positive_data_test.copy()
+    print(
+        f"Original number of positive feedbacks in test: {cleaned_positive_data_test.shape[0]}"
+    )
+    cleaned_positive_data_test = cleaned_positive_data_test.loc[
+        lambda df: df.user_id.apply(
+            lambda user_id: user_id in positive_data_train.user_id.values
+        )
+    ]
+    print(
+        f"Number of positive feedbacks in test after removing users not present in train: {cleaned_positive_data_test.shape[0]}"
+    )
+    cleaned_positive_data_test = cleaned_positive_data_test.loc[
+        lambda df: df.item_id.apply(
+            lambda item_id: item_id in positive_data_train.item_id.values
+        )
+    ]
+    print(
+        f"Number of positive feedbacks in test after removing offers not present in train: {cleaned_positive_data_test.shape[0]}"
+    )
+
+    # Get all offers the model may predict
+    all_item_ids = list(set(positive_data_train.item_id.values))
+    # Get all users in cleaned test data
+    all_test_user_ids = list(set(cleaned_positive_data_test.user_id.values))
+
+    hidden_items_number = 0
+    recommended_hidden_items_number = 0
+    prediction_number = 0
+    recommended_items = []
+
+    user_count = 0
+    unexpectedness = []
+    serendipity = []
+    new_types_ratio = []
+
+    for user_id in tqdm.tqdm(all_test_user_ids):
+        user_count += 1
+
+        positive_item_train = positive_data_train[
+            positive_data_train["user_id"] == user_id
+        ]
+        positive_item_test = cleaned_positive_data_test[
+            cleaned_positive_data_test["user_id"] == user_id
+        ]
+
+        # Remove items in train - they can not be in test set anyway
+        items_to_rank = np.setdiff1d(
+            all_item_ids, positive_item_train["item_id"].values
+        )
+        booked_offer_types = positive_item_train["type"].values
+
+        # Check if any item of items_to_rank is in the test positive feedbacks for this user
+        expected = np.in1d(items_to_rank, positive_item_test["item_id"].values)
+
+        repeated_user_id = np.empty_like(items_to_rank)
+        repeated_user_id.fill(user_id)
+
+        predicted = match_model.predict(
+            [repeated_user_id, items_to_rank], batch_size=4096
+        )
+
+        scored_items = sorted(
+            [(item_id, score[0]) for item_id, score in zip(items_to_rank, predicted)],
+            key=lambda k: k[1],
+            reverse=True,
+        )[:k]
+        recommended_offer_types = [offer_type_dict[item[0]] for item in scored_items]
+
+        user_unexpectedness = get_unexpectedness(
+            booked_offer_types, recommended_offer_types
+        )
+        unexpectedness.append(user_unexpectedness)
+        new_types_ratio.append(
+            np.mean(
+                [
+                    int(recommended_offer_type not in booked_offer_types)
+                    for recommended_offer_type in recommended_offer_types
+                ]
+            )
+        )
+
+        if np.sum(expected) >= 1:
+            recommended_items.extend([item[0] for item in scored_items])
+            recommended_items = list(set(recommended_items))
+
+            hidden_items = positive_item_test["item_id"].values
+            recommended_hidden_items = [
+                item[0] for item in scored_items if item[0] in hidden_items
+            ]
+
+            hidden_items_number += len(hidden_items)
+            recommended_hidden_items_number += len(recommended_hidden_items)
+            prediction_number += 1
+
+            if hidden_items_number > 0:
+                user_serendipity = (
+                    (len(recommended_hidden_items) / len(hidden_items))
+                    * user_unexpectedness
+                    * 100
+                )
+                serendipity.append(user_serendipity)
+
+        metrics = {
+            f"recall_at_{k}": (recommended_hidden_items_number / hidden_items_number)
+            * 100
+            if hidden_items_number > 0
+            else None,
+            f"precision_at_{k}": (
+                recommended_hidden_items_number / (prediction_number * k)
+            )
+            * 100,
+            f"maximal_precision_at_{k}": (
+                cleaned_positive_data_test.shape[0] / (prediction_number * k)
+            )
+            * 100,
+            f"coverage_at_{k}": (len(recommended_items) / len(all_item_ids)) * 100,
+            f"unexpectedness_at_{k}": np.mean(unexpectedness)
+            if len(unexpectedness) > 0
+            else None,
+            f"new_types_ratio_at_{k}": np.mean(new_types_ratio)
+            if len(new_types_ratio) > 0
+            else None,
+            f"serendipity_at_{k}": np.mean(serendipity)
+            if len(serendipity) > 0
+            else None,
+        }
+
+        if user_count % 100 == 0:
+            print(metrics)
+
+    return metrics
