+++ recommendation/api/recommendation.py
@@ -109,6 +109,7 @@ def get_final_recommendations(
     final_recommendations = order_offers_by_score_and_diversify_types(
         scored_recommendation_for_user, app_config["NUMBER_OF_RECOMMENDATIONS"]
     )
+
     save_recommendation(user_id, final_recommendations, connection)
     connection.close()
     return final_recommendations
